refactor(external-auth): 拆分外部认证服务与前端组件为独立模块
- 将 `external_auth_service.rs` 单文件拆分为 `mod.rs`、`links`、`login`、`normalize`、`password_link`、`providers`、`resolution`、`verification` 等子模块
- 将 `AdminExternalAuthPage.tsx` 中的逻辑提取为独立组件：`ExternalAuthCallbackDialog`、`ExternalAuthClaimFields`、`ExternalAuthConnectionFields`、`ExternalAuthCreateProgress`、`ExternalAuthProviderDialog`、`ExternalAuthProviderPanels`、`ExternalAuthProvidersTable` 及 `shared`
- 将 OIDC 集成测试从 `test_oidc.rs` 迁移至 `tests/external_auth/oidc/` 目录，新增 `dex.rs` 和 `mock.rs` 测试辅助模块
