+++ frontend-panel/src/components/admin/admin-external-auth-page/shared.tsx
@@ -1,0 +1,634 @@
+import type { TFunction } from "i18next";
+import type { MouseEvent } from "react";
+import { useTranslation } from "react-i18next";
+import { Button } from "@/components/ui/button";
+import { Icon } from "@/components/ui/icon";
+import {
+	externalAuthKindIconPath,
+	normalizeExternalAuthIconUrl,
+} from "@/lib/externalAuthProviders";
+import {
+	buildOffsetPaginationSearchParams,
+	parseOffsetSearchParam,
+	parsePageSizeSearchParam,
+} from "@/lib/pagination";
+import { absoluteAppUrl } from "@/lib/publicSiteUrl";
+import { cn } from "@/lib/utils";
+import type {
+	AdminExternalAuthProviderInfo,
+	AdminExternalAuthProviderKindInfo,
+	CreateExternalAuthProviderInput,
+	ExternalAuthProviderKind,
+	ExternalAuthProviderTestParamsInput,
+	ExternalAuthProviderTestResult,
+	UpdateExternalAuthProviderInput,
+} from "@/types/api";
+
+export const DEFAULT_SCOPES = "openid email profile";
+const REDACTED_SECRET = "***REDACTED***";
+export const STANDARD_CLAIMS = {
+	avatarUrlClaim: "picture",
+	displayNameClaim: "name",
+	emailClaim: "email",
+	emailVerifiedClaim: "email_verified",
+	groupsClaim: "groups",
+	subjectClaim: "sub",
+	usernameClaim: "preferred_username",
+} as const;
+export const EXTERNAL_AUTH_PAGE_SIZE_OPTIONS = [10, 20, 50] as const;
+export const DEFAULT_EXTERNAL_AUTH_PAGE_SIZE = 20 as const;
+const EXTERNAL_AUTH_MANAGED_QUERY_KEYS = ["offset", "pageSize"] as const;
+
+export interface ExternalAuthProviderFormData {
+	allowedDomains: string;
+	authorizationUrl: string;
+	autoLinkVerifiedEmailEnabled: boolean;
+	autoProvisionEnabled: boolean;
+	avatarUrlClaim: string;
+	clientId: string;
+	clientSecret: string;
+	displayName: string;
+	displayNameClaim: string;
+	emailClaim: string;
+	emailVerifiedClaim: string;
+	enabled: boolean;
+	groupsClaim: string;
+	iconUrl: string;
+	issuerUrl: string;
+	key: string;
+	providerKind: ExternalAuthProviderKind;
+	requireEmailVerified: boolean;
+	scopes: string;
+	subjectClaim: string;
+	tokenUrl: string;
+	userinfoUrl: string;
+	usernameClaim: string;
+}
+
+export type ExternalAuthProviderFieldChange = <
+	K extends keyof ExternalAuthProviderFormData,
+>(
+	key: K,
+	value: ExternalAuthProviderFormData[K],
+) => void;
+
+export interface ExternalAuthCreateStep {
+	title: string;
+	description: string;
+}
+
+export const emptyForm: ExternalAuthProviderFormData = {
+	allowedDomains: "",
+	authorizationUrl: "",
+	autoLinkVerifiedEmailEnabled: false,
+	autoProvisionEnabled: false,
+	avatarUrlClaim: "",
+	clientId: "",
+	clientSecret: "",
+	displayName: "",
+	displayNameClaim: "",
+	emailClaim: "",
+	emailVerifiedClaim: "",
+	enabled: true,
+	groupsClaim: "",
+	iconUrl: "",
+	issuerUrl: "",
+	key: "",
+	providerKind: "oidc",
+	requireEmailVerified: true,
+	scopes: DEFAULT_SCOPES,
+	subjectClaim: "",
+	tokenUrl: "",
+	userinfoUrl: "",
+	usernameClaim: "",
+};
+
+export function formFromProvider(
+	provider: AdminExternalAuthProviderInfo,
+): ExternalAuthProviderFormData {
+	return {
+		allowedDomains: provider.allowed_domains.join(", "),
+		authorizationUrl: provider.authorization_url ?? "",
+		autoLinkVerifiedEmailEnabled: provider.auto_link_verified_email_enabled,
+		autoProvisionEnabled: provider.auto_provision_enabled,
+		avatarUrlClaim: provider.avatar_url_claim ?? "",
+		clientId: provider.client_id,
+		clientSecret: provider.client_secret ?? "",
+		displayName: provider.display_name,
+		displayNameClaim: provider.display_name_claim ?? "",
+		emailClaim: provider.email_claim ?? "",
+		emailVerifiedClaim: provider.email_verified_claim ?? "",
+		enabled: provider.enabled,
+		groupsClaim: provider.groups_claim ?? "",
+		iconUrl: provider.icon_url ?? "",
+		issuerUrl: provider.issuer_url ?? "",
+		key: provider.key,
+		providerKind: provider.provider_kind,
+		requireEmailVerified: provider.require_email_verified,
+		scopes: provider.scopes || DEFAULT_SCOPES,
+		subjectClaim: provider.subject_claim ?? "",
+		tokenUrl: provider.token_url ?? "",
+		userinfoUrl: provider.userinfo_url ?? "",
+		usernameClaim: provider.username_claim ?? "",
+	};
+}
+
+function kindFallbackLabel(kind: ExternalAuthProviderKind) {
+	switch (kind) {
+		case "oidc":
+			return "OpenID Connect";
+	}
+}
+
+function localizedProviderKindText(
+	t: TFunction,
+	key: string,
+	fallback: string,
+) {
+	const translated = t(key);
+	return translated === key ? fallback : translated;
+}
+
+export function ExternalAuthProviderIcon({
+	className,
+	iconUrl,
+	kind,
+}: {
+	className?: string;
+	iconUrl?: string | null;
+	kind: ExternalAuthProviderKind;
+}) {
+	const configuredIcon = normalizeExternalAuthIconUrl(iconUrl);
+	const kindIcon = externalAuthKindIconPath(kind);
+	const effectiveIcon = configuredIcon || kindIcon;
+
+	if (effectiveIcon) {
+		return (
+			<img
+				src={effectiveIcon}
+				alt=""
+				aria-hidden="true"
+				className={cn("object-contain", className)}
+				onError={(event) => {
+					if (
+						configuredIcon &&
+						kindIcon &&
+						event.currentTarget.src !== kindIcon
+					) {
+						event.currentTarget.src = kindIcon;
+						return;
+					}
+					event.currentTarget.hidden = true;
+				}}
+			/>
+		);
+	}
+
+	return <Icon name="SignIn" className={cn("text-primary", className)} />;
+}
+
+export function kindDisplayName(
+	t: TFunction,
+	kind: ExternalAuthProviderKind,
+	providerKinds: AdminExternalAuthProviderKindInfo[],
+) {
+	const fallback =
+		providerKinds.find((item) => item.kind === kind)?.display_name ??
+		kindFallbackLabel(kind);
+	return localizedProviderKindText(
+		t,
+		`external_auth_provider_kind_${kind}_name`,
+		fallback,
+	);
+}
+
+export function kindDescription(
+	t: TFunction,
+	kind: AdminExternalAuthProviderKindInfo,
+) {
+	return localizedProviderKindText(
+		t,
+		`external_auth_provider_kind_${kind.kind}_description`,
+		kind.description,
+	);
+}
+
+export function parseAllowedDomains(value: string) {
+	return value
+		.split(/[,\n]/)
+		.map((domain) => domain.trim().replace(/^@+/, "").toLowerCase())
+		.filter(
+			(domain, index, domains) => domain && domains.indexOf(domain) === index,
+		);
+}
+
+function nullableText(value: string) {
+	const trimmed = value.trim();
+	return trimmed ? trimmed : null;
+}
+
+function nullableSecretText(value: string) {
+	const trimmed = value.trim();
+	return trimmed && trimmed !== REDACTED_SECRET ? trimmed : null;
+}
+
+function effectiveClaim(value: string | null | undefined, fallback: string) {
+	return value?.trim() || fallback;
+}
+
+export function createPayload(
+	form: ExternalAuthProviderFormData,
+): CreateExternalAuthProviderInput {
+	const allowedDomains = parseAllowedDomains(form.allowedDomains);
+	return {
+		allowed_domains: allowedDomains.length > 0 ? allowedDomains : null,
+		authorization_url: nullableText(form.authorizationUrl),
+		auto_link_verified_email_enabled: form.autoLinkVerifiedEmailEnabled,
+		auto_provision_enabled: form.autoProvisionEnabled,
+		avatar_url_claim: nullableText(form.avatarUrlClaim),
+		client_id: form.clientId.trim(),
+		client_secret: nullableText(form.clientSecret),
+		display_name: form.displayName.trim(),
+		display_name_claim: nullableText(form.displayNameClaim),
+		email_claim: nullableText(form.emailClaim),
+		email_verified_claim: nullableText(form.emailVerifiedClaim),
+		enabled: form.enabled,
+		groups_claim: nullableText(form.groupsClaim),
+		icon_url: nullableText(form.iconUrl),
+		issuer_url: nullableText(form.issuerUrl),
+		provider_kind: form.providerKind,
+		require_email_verified: form.requireEmailVerified,
+		scopes: form.scopes.trim() || DEFAULT_SCOPES,
+		subject_claim: nullableText(form.subjectClaim),
+		token_url: nullableText(form.tokenUrl),
+		userinfo_url: nullableText(form.userinfoUrl),
+		username_claim: nullableText(form.usernameClaim),
+	};
+}
+
+export function updatePayload(
+	form: ExternalAuthProviderFormData,
+): UpdateExternalAuthProviderInput {
+	const allowedDomains = parseAllowedDomains(form.allowedDomains);
+	return {
+		allowed_domains: allowedDomains.length > 0 ? allowedDomains : null,
+		authorization_url: nullableText(form.authorizationUrl),
+		auto_link_verified_email_enabled: form.autoLinkVerifiedEmailEnabled,
+		auto_provision_enabled: form.autoProvisionEnabled,
+		avatar_url_claim: nullableText(form.avatarUrlClaim),
+		client_id: form.clientId.trim(),
+		client_secret: nullableText(form.clientSecret),
+		display_name: form.displayName.trim(),
+		display_name_claim: nullableText(form.displayNameClaim),
+		email_claim: nullableText(form.emailClaim),
+		email_verified_claim: nullableText(form.emailVerifiedClaim),
+		enabled: form.enabled,
+		groups_claim: nullableText(form.groupsClaim),
+		icon_url: nullableText(form.iconUrl),
+		issuer_url: nullableText(form.issuerUrl),
+		require_email_verified: form.requireEmailVerified,
+		scopes: form.scopes.trim() || DEFAULT_SCOPES,
+		subject_claim: nullableText(form.subjectClaim),
+		token_url: nullableText(form.tokenUrl),
+		userinfo_url: nullableText(form.userinfoUrl),
+		username_claim: nullableText(form.usernameClaim),
+	};
+}
+
+export function testParamsPayload(
+	form: ExternalAuthProviderFormData,
+): ExternalAuthProviderTestParamsInput {
+	return {
+		authorization_url: nullableText(form.authorizationUrl),
+		client_id: form.clientId.trim(),
+		client_secret: nullableSecretText(form.clientSecret),
+		issuer_url: nullableText(form.issuerUrl),
+		provider_kind: form.providerKind,
+		scopes: form.scopes.trim() || DEFAULT_SCOPES,
+		token_url: nullableText(form.tokenUrl),
+		userinfo_url: nullableText(form.userinfoUrl),
+	};
+}
+
+function normalizeConnectionValue(value: string | null | undefined) {
+	return value?.trim() ?? "";
+}
+
+function formClientSecretChanged(
+	form: ExternalAuthProviderFormData,
+	provider: AdminExternalAuthProviderInfo,
+) {
+	const value = normalizeConnectionValue(form.clientSecret);
+	return provider.client_secret_configured
+		? value !== REDACTED_SECRET
+		: value !== "";
+}
+
+export function formConnectionChanged(
+	form: ExternalAuthProviderFormData,
+	provider: AdminExternalAuthProviderInfo,
+) {
+	return (
+		form.providerKind !== provider.provider_kind ||
+		normalizeConnectionValue(form.issuerUrl) !==
+			normalizeConnectionValue(provider.issuer_url) ||
+		normalizeConnectionValue(form.authorizationUrl) !==
+			normalizeConnectionValue(provider.authorization_url) ||
+		normalizeConnectionValue(form.tokenUrl) !==
+			normalizeConnectionValue(provider.token_url) ||
+		normalizeConnectionValue(form.userinfoUrl) !==
+			normalizeConnectionValue(provider.userinfo_url) ||
+		normalizeConnectionValue(form.clientId) !==
+			normalizeConnectionValue(provider.client_id) ||
+		(form.scopes.trim() || DEFAULT_SCOPES) !==
+			(provider.scopes.trim() || DEFAULT_SCOPES) ||
+		formClientSecretChanged(form, provider)
+	);
+}
+
+export function formatTestResultSummary(
+	t: TFunction,
+	result: ExternalAuthProviderTestResult,
+) {
+	return result.checks.length > 0
+		? result.checks
+				.map((check) =>
+					t(
+						check.success
+							? "external_auth_provider_test_check_ok"
+							: "external_auth_provider_test_check_error",
+						{
+							name: check.name,
+							message: check.message,
+						},
+					),
+				)
+				.join(" · ")
+		: t("external_auth_provider_test_success_detail", {
+				provider: result.provider,
+			});
+}
+
+export function providerStatusTone(provider: AdminExternalAuthProviderInfo) {
+	return provider.enabled
+		? "border-emerald-200 bg-emerald-50 text-emerald-700 dark:border-emerald-900 dark:bg-emerald-950/60 dark:text-emerald-300"
+		: "border-slate-200 bg-slate-50 text-slate-700 dark:border-slate-800 dark:bg-slate-950/50 dark:text-slate-300";
+}
+
+export function securityModeLabel(
+	t: TFunction,
+	provider: AdminExternalAuthProviderInfo,
+) {
+	if (
+		provider.auto_provision_enabled &&
+		provider.auto_link_verified_email_enabled
+	) {
+		return t("external_auth_provider_mode_link_and_provision");
+	}
+	if (provider.auto_provision_enabled) {
+		return t("external_auth_provider_mode_provision");
+	}
+	if (provider.auto_link_verified_email_enabled) {
+		return t("external_auth_provider_mode_link");
+	}
+	return t("external_auth_provider_mode_manual");
+}
+
+function callbackPath(
+	providerKind: ExternalAuthProviderKind,
+	providerKey: string,
+) {
+	const key = providerKey.trim();
+	return key
+		? `/api/v1/auth/external-auth/${encodeURIComponent(providerKind)}/${encodeURIComponent(key)}/callback`
+		: null;
+}
+
+export function callbackUrl(
+	providerKind: ExternalAuthProviderKind,
+	providerKey: string,
+) {
+	const path = callbackPath(providerKind, providerKey);
+	return path ? absoluteAppUrl(path) : "";
+}
+
+export function providerPrimaryEndpoint(
+	provider: AdminExternalAuthProviderInfo,
+) {
+	if (provider.issuer_url) {
+		return {
+			labelKey: "external_auth_provider_issuer_url",
+			value: provider.issuer_url,
+		};
+	}
+	if (provider.authorization_url) {
+		return {
+			labelKey: "external_auth_provider_authorization_url",
+			value: provider.authorization_url,
+		};
+	}
+	if (provider.token_url) {
+		return {
+			labelKey: "external_auth_provider_token_url",
+			value: provider.token_url,
+		};
+	}
+	if (provider.userinfo_url) {
+		return {
+			labelKey: "external_auth_provider_userinfo_url",
+			value: provider.userinfo_url,
+		};
+	}
+	return null;
+}
+
+export function providerAllowedDomainSummary(
+	t: TFunction,
+	provider: AdminExternalAuthProviderInfo,
+) {
+	return provider.allowed_domains.length > 0
+		? provider.allowed_domains.join(", ")
+		: t("external_auth_provider_allowed_domains_all");
+}
+
+export function normalizeOffset(offset: number) {
+	return Math.max(0, Math.floor(offset));
+}
+
+export function buildManagedExternalAuthSearchParams({
+	offset,
+	pageSize,
+}: {
+	offset: number;
+	pageSize: (typeof EXTERNAL_AUTH_PAGE_SIZE_OPTIONS)[number];
+}) {
+	return buildOffsetPaginationSearchParams({
+		offset,
+		pageSize,
+		defaultPageSize: DEFAULT_EXTERNAL_AUTH_PAGE_SIZE,
+	});
+}
+
+export function getManagedExternalAuthSearchString(
+	searchParams: URLSearchParams,
+) {
+	return buildManagedExternalAuthSearchParams({
+		offset: normalizeOffset(parseOffsetSearchParam(searchParams.get("offset"))),
+		pageSize: parsePageSizeSearchParam(
+			searchParams.get("pageSize"),
+			EXTERNAL_AUTH_PAGE_SIZE_OPTIONS,
+			DEFAULT_EXTERNAL_AUTH_PAGE_SIZE,
+		),
+	}).toString();
+}
+
+export function mergeManagedExternalAuthSearchParams(
+	searchParams: URLSearchParams,
+	managedSearchParams: URLSearchParams,
+) {
+	const merged = new URLSearchParams(searchParams);
+	for (const key of EXTERNAL_AUTH_MANAGED_QUERY_KEYS) {
+		merged.delete(key);
+	}
+	for (const [key, value] of managedSearchParams.entries()) {
+		merged.set(key, value);
+	}
+	return merged;
+}
+
+export function shouldShowIssuerUrl(
+	kind: AdminExternalAuthProviderKindInfo | null,
+) {
+	return Boolean(kind?.supports_discovery || kind?.issuer_url_required);
+}
+
+export function shouldShowManualEndpoints(
+	kind: AdminExternalAuthProviderKindInfo | null,
+) {
+	return Boolean(kind?.manual_endpoint_configuration_supported);
+}
+
+export function connectionRequirementsMissing(
+	form: ExternalAuthProviderFormData,
+	kind: AdminExternalAuthProviderKindInfo | null,
+) {
+	if (!form.clientId.trim()) {
+		return true;
+	}
+	if ((kind?.issuer_url_required ?? true) && !form.issuerUrl.trim()) {
+		return true;
+	}
+	if (kind?.authorization_url_required && !form.authorizationUrl.trim()) {
+		return true;
+	}
+	if (kind?.token_url_required && !form.tokenUrl.trim()) {
+		return true;
+	}
+	if (kind?.userinfo_url_required && !form.userinfoUrl.trim()) {
+		return true;
+	}
+	return false;
+}
+
+export function requiredFieldsMissing(
+	form: ExternalAuthProviderFormData,
+	kind: AdminExternalAuthProviderKindInfo | null,
+) {
+	return !form.displayName.trim() || connectionRequirementsMissing(form, kind);
+}
+
+export function formConnectionSummary(
+	form: ExternalAuthProviderFormData,
+	selectedKind: AdminExternalAuthProviderKindInfo | null,
+) {
+	const items = [
+		form.issuerUrl.trim() ? `issuer: ${form.issuerUrl.trim()}` : null,
+		selectedKind?.manual_endpoint_configuration_supported &&
+		form.authorizationUrl.trim()
+			? `authorization: ${form.authorizationUrl.trim()}`
+			: null,
+		selectedKind?.manual_endpoint_configuration_supported &&
+		form.tokenUrl.trim()
+			? `token: ${form.tokenUrl.trim()}`
+			: null,
+		selectedKind?.manual_endpoint_configuration_supported &&
+		form.userinfoUrl.trim()
+			? `userinfo: ${form.userinfoUrl.trim()}`
+			: null,
+	]
+		.filter((item): item is string => item !== null)
+		.join(" · ");
+	return items || "-";
+}
+
+export function formClaimSummary(
+	form: ExternalAuthProviderFormData,
+	selectedKind: AdminExternalAuthProviderKindInfo | null,
+) {
+	const claims = [
+		`subject=${effectiveClaim(form.subjectClaim, STANDARD_CLAIMS.subjectClaim)}`,
+		`username=${effectiveClaim(form.usernameClaim, STANDARD_CLAIMS.usernameClaim)}`,
+		`display=${effectiveClaim(form.displayNameClaim, STANDARD_CLAIMS.displayNameClaim)}`,
+		`email=${effectiveClaim(form.emailClaim, STANDARD_CLAIMS.emailClaim)}`,
+		selectedKind?.supports_email_verified_claim
+			? `email_verified=${effectiveClaim(form.emailVerifiedClaim, STANDARD_CLAIMS.emailVerifiedClaim)}`
+			: null,
+		`groups=${effectiveClaim(form.groupsClaim, STANDARD_CLAIMS.groupsClaim)}`,
+		`avatar=${effectiveClaim(form.avatarUrlClaim, STANDARD_CLAIMS.avatarUrlClaim)}`,
+	]
+		.filter((item): item is string => item !== null)
+		.join(" · ");
+	return claims || "-";
+}
+
+export function providerIconSummary(form: ExternalAuthProviderFormData) {
+	return form.iconUrl.trim() || "-";
+}
+
+interface CallbackUrlFieldProps {
+	className?: string;
+	onCopy: (value: string) => void;
+	value: string;
+}
+
+export function CallbackUrlField({
+	className,
+	onCopy,
+	value,
+}: CallbackUrlFieldProps) {
+	const { t } = useTranslation("admin");
+	const disabled = !value;
+
+	return (
+		<div
+			className={cn(
+				"flex min-w-0 w-full max-w-full items-center gap-2 overflow-hidden rounded-md border border-border/70 bg-muted/30 p-1",
+				className,
+			)}
+		>
+			<code
+				className="block min-w-0 flex-1 select-all overflow-x-auto whitespace-nowrap px-2 py-1 font-mono text-xs text-foreground [scrollbar-width:thin]"
+				title={value || "-"}
+			>
+				{value || "-"}
+			</code>
+			<Button
+				type="button"
+				variant="ghost"
+				size="icon"
+				className="h-7 w-7 shrink-0"
+				disabled={disabled}
+				aria-label={t("external_auth_provider_copy_callback_url")}
+				title={t("external_auth_provider_copy_callback_url")}
+				onClick={(event: MouseEvent<HTMLButtonElement>) => {
+					event.stopPropagation();
+					if (!disabled) {
+						onCopy(value);
+					}
+				}}
+			>
+				<Icon name="Copy" className="h-3.5 w-3.5" />
+			</Button>
+		</div>
+	);
+}
