+++ tests/external_auth/mod.rs
@@ -1,0 +1,1 @@
+pub mod oidc;
