+++ tests/test_oidc.rs
@@ -3,9 +3,9 @@
 #[macro_use]
 mod common;
 
-use actix_web::{
-    App, HttpResponse, HttpServer, Responder, body::MessageBody, dev::ServiceResponse, test, web,
-};
+mod external_auth;
+
+use actix_web::test;
 use aster_drive::db::repository::{
     external_auth_identity_repo, external_auth_login_flow_repo, external_auth_provider_repo,
 };
@@ -14,864 +14,13 @@ use aster_drive::entities::{
     external_auth_provider, user,
 };
 use aster_drive::services::external_auth_service;
-use base64::Engine as _;
 use chrono::{Duration, Utc};
-use jsonwebtoken::{
-    Algorithm, EncodingKey, Header,
-    jwk::{
-        AlgorithmParameters, CommonParameters, Jwk, JwkSet, KeyAlgorithm, PublicKeyUse,
-        RSAKeyParameters,
-    },
-};
-use rsa::{
-    RsaPrivateKey, RsaPublicKey, pkcs1::EncodeRsaPrivateKey, rand_core::OsRng,
-    traits::PublicKeyParts,
-};
+use external_auth::oidc::*;
 use sea_orm::{
     ActiveModelTrait, ActiveValue::Set, ColumnTrait, EntityTrait, IntoActiveModel, QueryFilter,
 };
-use serde::Deserialize;
 use serde_json::Value;
-use std::sync::{Arc, Mutex};
-use std::time::Duration as StdDuration;
 use uuid::Uuid;
-
-const TEST_BROWSER_ORIGIN: &str = "http://localhost:8080";
-const TEST_CLIENT_ID: &str = "aster-test-client";
-const TEST_KID: &str = "aster-test-kid";
-const DEX_TEST_CLIENT_SECRET: &str = "super-secret";
-const DEX_TEST_IMAGE_TAG: &str = "v2.42.0";
-const DEX_TEST_USER_EMAIL: &str = "dex-user@example.com";
-const DEX_TEST_USER_SUBJECT: &str = "CgtkZXgtdXNlci1pZBIFbG9jYWw";
-
-#[derive(Clone)]
-struct MockOidcProvider {
-    issuer: String,
-    key: Arc<RsaPrivateKey>,
-    authorization_requests: Arc<Mutex<Vec<AuthorizeRequest>>>,
-    token_subject: Arc<Mutex<String>>,
-    token_email: Arc<Mutex<Option<String>>>,
-    token_email_verified: Arc<Mutex<bool>>,
-    token_audience: Arc<Mutex<String>>,
-    token_nonce_override: Arc<Mutex<Option<String>>>,
-    token_issuer_override: Arc<Mutex<Option<String>>>,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-struct AuthorizeRequest {
-    response_type: String,
-    client_id: String,
-    redirect_uri: String,
-    scope: Option<String>,
-    state: String,
-    nonce: String,
-    code_challenge: Option<String>,
-    code_challenge_method: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct TokenRequest {
-    grant_type: String,
-    code: String,
-    redirect_uri: String,
-    client_id: Option<String>,
-    client_secret: Option<String>,
-    code_verifier: Option<String>,
-}
-
-impl MockOidcProvider {
-    fn new() -> Self {
-        let mut rng = OsRng;
-        let key = RsaPrivateKey::new(&mut rng, 2048).expect("RSA key should generate");
-        Self {
-            issuer: String::new(),
-            key: Arc::new(key),
-            authorization_requests: Arc::new(Mutex::new(Vec::new())),
-            token_subject: Arc::new(Mutex::new("oidc-subject-1".to_string())),
-            token_email: Arc::new(Mutex::new(Some("oidc-user@example.com".to_string()))),
-            token_email_verified: Arc::new(Mutex::new(true)),
-            token_audience: Arc::new(Mutex::new(TEST_CLIENT_ID.to_string())),
-            token_nonce_override: Arc::new(Mutex::new(None)),
-            token_issuer_override: Arc::new(Mutex::new(None)),
-        }
-    }
-
-    fn with_issuer(mut self, issuer: String) -> Self {
-        self.issuer = issuer;
-        self
-    }
-
-    fn last_authorize_request(&self) -> AuthorizeRequest {
-        self.authorization_requests
-            .lock()
-            .expect("authorize requests lock should not be poisoned")
-            .last()
-            .expect("authorization request should be recorded")
-            .clone()
-    }
-
-    fn set_issuer_override(&self, issuer: Option<String>) {
-        *self
-            .token_issuer_override
-            .lock()
-            .expect("issuer override lock should not be poisoned") = issuer;
-    }
-
-    fn set_subject(&self, subject: &str) {
-        *self
-            .token_subject
-            .lock()
-            .expect("subject lock should not be poisoned") = subject.to_string();
-    }
-
-    fn set_email(&self, email: &str) {
-        *self
-            .token_email
-            .lock()
-            .expect("email lock should not be poisoned") = Some(email.to_string());
-    }
-
-    fn clear_email(&self) {
-        *self
-            .token_email
-            .lock()
-            .expect("email lock should not be poisoned") = None;
-    }
-
-    fn set_email_verified(&self, verified: bool) {
-        *self
-            .token_email_verified
-            .lock()
-            .expect("email verified lock should not be poisoned") = verified;
-    }
-
-    fn set_audience(&self, audience: &str) {
-        *self
-            .token_audience
-            .lock()
-            .expect("audience lock should not be poisoned") = audience.to_string();
-    }
-
-    fn set_nonce_override(&self, nonce: Option<String>) {
-        *self
-            .token_nonce_override
-            .lock()
-            .expect("nonce override lock should not be poisoned") = nonce;
-    }
-
-    fn public_jwk(&self) -> Jwk {
-        let public = RsaPublicKey::from(self.key.as_ref());
-        Jwk {
-            common: CommonParameters {
-                public_key_use: Some(PublicKeyUse::Signature),
-                key_algorithm: Some(KeyAlgorithm::RS256),
-                key_id: Some(TEST_KID.to_string()),
-                ..Default::default()
-            },
-            algorithm: AlgorithmParameters::RSA(RSAKeyParameters {
-                n: base64_url(public.n().to_bytes_be()),
-                e: base64_url(public.e().to_bytes_be()),
-                ..Default::default()
-            }),
-        }
-    }
-
-    fn sign_id_token(&self, nonce: &str) -> String {
-        let issuer_override = self
-            .token_issuer_override
-            .lock()
-            .expect("issuer override lock should not be poisoned")
-            .clone();
-        let issuer = issuer_override.as_deref().unwrap_or(&self.issuer);
-        let subject = self
-            .token_subject
-            .lock()
-            .expect("subject lock should not be poisoned")
-            .clone();
-        let email = self
-            .token_email
-            .lock()
-            .expect("email lock should not be poisoned")
-            .clone();
-        let email_verified = *self
-            .token_email_verified
-            .lock()
-            .expect("email verified lock should not be poisoned");
-        let audience = self
-            .token_audience
-            .lock()
-            .expect("audience lock should not be poisoned")
-            .clone();
-        let nonce_override = self
-            .token_nonce_override
-            .lock()
-            .expect("nonce override lock should not be poisoned")
-            .clone();
-        let nonce = nonce_override.as_deref().unwrap_or(nonce);
-        let now = Utc::now();
-        let mut claims = serde_json::json!({
-            "iss": issuer,
-            "sub": subject,
-            "aud": audience,
-            "exp": (now + Duration::minutes(5)).timestamp(),
-            "iat": now.timestamp(),
-            "nonce": nonce,
-            "name": "OIDC Test User",
-            "preferred_username": "oidctest"
-        });
-        if let Some(email) = email {
-            claims["email"] = serde_json::json!(email);
-            claims["email_verified"] = serde_json::json!(email_verified);
-        }
-        let mut header = Header::new(Algorithm::RS256);
-        header.kid = Some(TEST_KID.to_string());
-        let der = self
-            .key
-            .to_pkcs1_der()
-            .expect("private key should encode to pkcs1");
-        jsonwebtoken::encode(&header, &claims, &EncodingKey::from_rsa_der(der.as_bytes()))
-            .expect("id_token should sign")
-    }
-}
-
-fn base64_url(bytes: Vec<u8>) -> String {
-    base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(bytes)
-}
-
-async fn start_mock_external_auth_provider() -> (MockOidcProvider, actix_web::dev::ServerHandle) {
-    let seed = MockOidcProvider::new();
-    let listener = std::net::TcpListener::bind(("127.0.0.1", 0)).expect("listener should bind");
-    let addr = listener
-        .local_addr()
-        .expect("listener address should exist");
-    let provider = seed.with_issuer(format!(
-        "http://127.0.0.1:{addr_port}",
-        addr_port = addr.port()
-    ));
-    let app_provider = provider.clone();
-    let server = HttpServer::new(move || {
-        App::new()
-            .app_data(web::Data::new(app_provider.clone()))
-            .route(
-                "/.well-known/openid-configuration",
-                web::get().to(mock_discovery),
-            )
-            .route("/authorize", web::get().to(mock_authorize))
-            .route("/token", web::post().to(mock_token))
-            .route("/jwks", web::get().to(mock_jwks))
-    })
-    .listen(listener)
-    .expect("mock OIDC server should listen")
-    .run();
-    let handle = server.handle();
-    tokio::spawn(server);
-    (provider, handle)
-}
-
-async fn mock_discovery(provider: web::Data<MockOidcProvider>) -> impl Responder {
-    HttpResponse::Ok().json(serde_json::json!({
-        "issuer": provider.issuer,
-        "authorization_endpoint": format!("{}/authorize", provider.issuer),
-        "token_endpoint": format!("{}/token", provider.issuer),
-        "jwks_uri": format!("{}/jwks", provider.issuer),
-        "response_types_supported": ["code"],
-        "subject_types_supported": ["public"],
-        "id_token_signing_alg_values_supported": ["RS256"],
-        "scopes_supported": ["openid", "email", "profile"],
-        "token_endpoint_auth_methods_supported": ["none", "client_secret_post", "client_secret_basic"],
-        "claims_supported": ["sub", "iss", "aud", "exp", "iat", "nonce", "email", "email_verified", "name", "preferred_username"],
-        "code_challenge_methods_supported": ["S256"]
-    }))
-}
-
-async fn mock_authorize(
-    provider: web::Data<MockOidcProvider>,
-    query: web::Query<AuthorizeRequest>,
-) -> impl Responder {
-    provider
-        .authorization_requests
-        .lock()
-        .expect("authorize requests lock should not be poisoned")
-        .push(query.into_inner());
-    HttpResponse::Ok().finish()
-}
-
-async fn mock_token(
-    provider: web::Data<MockOidcProvider>,
-    form: web::Form<TokenRequest>,
-) -> impl Responder {
-    let request = form.into_inner();
-    assert_eq!(request.grant_type, "authorization_code");
-    assert_eq!(request.code, "mock-code");
-    if let Some(client_id) = request.client_id.as_deref() {
-        assert_eq!(client_id, TEST_CLIENT_ID);
-    }
-    if let Some(client_secret) = request.client_secret.as_deref() {
-        assert_eq!(client_secret, "super-secret");
-    }
-    assert!(!request.redirect_uri.is_empty());
-    assert!(
-        request
-            .code_verifier
-            .as_deref()
-            .is_some_and(|value| !value.is_empty()),
-        "PKCE code_verifier should be sent to token endpoint"
-    );
-    let nonce = provider.last_authorize_request().nonce;
-    HttpResponse::Ok().json(serde_json::json!({
-        "access_token": "mock-access-token",
-        "token_type": "Bearer",
-        "expires_in": 300,
-        "id_token": provider.sign_id_token(&nonce)
-    }))
-}
-
-async fn mock_jwks(provider: web::Data<MockOidcProvider>) -> impl Responder {
-    HttpResponse::Ok().json(JwkSet {
-        keys: vec![provider.public_jwk()],
-    })
-}
-
-async fn create_external_auth_provider<S, B, E>(
-    app: &S,
-    admin_token: &str,
-    issuer_url: &str,
-    enabled: bool,
-    auto_provision_enabled: bool,
-) -> Value
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let mut options = TestOidcProviderOptions::mock(issuer_url);
-    options.enabled = enabled;
-    options.auto_provision_enabled = auto_provision_enabled;
-    create_external_auth_provider_with(app, admin_token, options).await
-}
-
-async fn create_external_auth_provider_key<S, B, E>(
-    app: &S,
-    admin_token: &str,
-    issuer_url: &str,
-    enabled: bool,
-    auto_provision_enabled: bool,
-) -> String
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let created = create_external_auth_provider(
-        app,
-        admin_token,
-        issuer_url,
-        enabled,
-        auto_provision_enabled,
-    )
-    .await;
-    created_provider_key(&created)
-}
-
-async fn create_external_auth_provider_with_key<S, B, E>(
-    app: &S,
-    admin_token: &str,
-    options: TestOidcProviderOptions,
-) -> String
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let created = create_external_auth_provider_with(app, admin_token, options).await;
-    created_provider_key(&created)
-}
-
-struct TestOidcProviderOptions {
-    display_name_prefix: String,
-    issuer_url: String,
-    enabled: bool,
-    auto_provision_enabled: bool,
-    auto_link_verified_email_enabled: bool,
-    require_email_verified: bool,
-    allowed_domains: Vec<String>,
-}
-
-impl TestOidcProviderOptions {
-    fn mock(issuer_url: &str) -> Self {
-        Self {
-            display_name_prefix: "mock".to_string(),
-            issuer_url: issuer_url.to_string(),
-            enabled: true,
-            auto_provision_enabled: false,
-            auto_link_verified_email_enabled: false,
-            require_email_verified: true,
-            allowed_domains: vec!["example.com".to_string()],
-        }
-    }
-}
-
-async fn create_external_auth_provider_with<S, B, E>(
-    app: &S,
-    admin_token: &str,
-    options: TestOidcProviderOptions,
-) -> Value
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let req = test::TestRequest::post()
-        .uri("/api/v1/admin/external-auth/providers")
-        .insert_header(("Cookie", common::access_cookie_header(admin_token)))
-        .insert_header(common::csrf_header_for(admin_token))
-        .set_json(serde_json::json!({
-            "provider_kind": "oidc",
-            "display_name": format!("{} OIDC", options.display_name_prefix),
-            "icon_url": "/static/external-auth/mock.svg",
-            "issuer_url": options.issuer_url,
-            "client_id": TEST_CLIENT_ID,
-            "client_secret": "super-secret",
-            "enabled": options.enabled,
-            "auto_provision_enabled": options.auto_provision_enabled,
-            "auto_link_verified_email_enabled": options.auto_link_verified_email_enabled,
-            "require_email_verified": options.require_email_verified,
-            "allowed_domains": options.allowed_domains
-        }))
-        .to_request();
-    let resp = test::call_service(app, req).await;
-    assert_eq!(resp.status(), 201);
-    test::read_body_json(resp).await
-}
-
-fn created_provider_key(created: &Value) -> String {
-    created["data"]["key"]
-        .as_str()
-        .expect("provider key should be returned")
-        .to_string()
-}
-
-fn external_auth_provider_model(
-    key: &str,
-    issuer_url: &str,
-    enabled: bool,
-) -> external_auth_provider::ActiveModel {
-    let now = Utc::now();
-    external_auth_provider::ActiveModel {
-        key: Set(key.to_string()),
-        display_name: Set(format!("{key} provider")),
-        icon_url: Set(None),
-        provider_kind: Set(aster_drive::types::ExternalAuthProviderKind::Oidc),
-        protocol: Set(aster_drive::types::ExternalAuthProtocol::Oidc),
-        issuer_url: Set(Some(issuer_url.to_string())),
-        authorization_url: Set(None),
-        token_url: Set(None),
-        userinfo_url: Set(None),
-        client_id: Set(TEST_CLIENT_ID.to_string()),
-        client_secret: Set(None),
-        scopes: Set("openid email profile".to_string()),
-        enabled: Set(enabled),
-        auto_provision_enabled: Set(false),
-        auto_link_verified_email_enabled: Set(false),
-        require_email_verified: Set(true),
-        subject_claim: Set(None),
-        username_claim: Set(None),
-        display_name_claim: Set(None),
-        email_claim: Set(None),
-        email_verified_claim: Set(None),
-        groups_claim: Set(None),
-        avatar_url_claim: Set(None),
-        allowed_domains: Set(None),
-        created_at: Set(now),
-        updated_at: Set(now),
-        ..Default::default()
-    }
-}
-
-fn configure_oidc_public_site_url(state: &aster_drive::runtime::PrimaryAppState) {
-    state.runtime_config.apply(common::system_config_model(
-        aster_drive::config::site_url::PUBLIC_SITE_URL_KEY,
-        r#"["http://localhost:8080"]"#,
-    ));
-}
-
-async fn start_oidc_login<S, B, E>(
-    app: &S,
-    mock_provider: &MockOidcProvider,
-    provider_key: &str,
-    return_path: &str,
-) -> String
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let req = test::TestRequest::post()
-        .uri(&format!(
-            "/api/v1/auth/external-auth/oidc/{provider_key}/start"
-        ))
-        .insert_header(("Origin", TEST_BROWSER_ORIGIN))
-        .set_json(serde_json::json!({ "return_path": return_path }))
-        .to_request();
-    let resp = test::call_service(app, req).await;
-    assert_eq!(resp.status(), 200);
-    let body: Value = test::read_body_json(resp).await;
-    let auth_url = body["data"]["authorization_url"]
-        .as_str()
-        .expect("authorization url should be returned");
-    reqwest::get(auth_url)
-        .await
-        .expect("mock authorize request should succeed");
-    mock_provider.last_authorize_request().state
-}
-
-async fn finish_oidc_callback<S, B, E>(
-    app: &S,
-    provider_key: &str,
-    state_value: &str,
-) -> ServiceResponse<B>
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let callback = format!(
-        "/api/v1/auth/external-auth/oidc/{provider_key}/callback?code=mock-code&state={state_value}"
-    );
-    let req = test::TestRequest::get()
-        .uri(&callback)
-        .peer_addr("127.0.0.1:12345".parse().unwrap())
-        .to_request();
-    test::call_service(app, req).await
-}
-
-fn assert_oidc_error_redirect<B>(resp: &ServiceResponse<B>) {
-    assert_eq!(resp.status(), 302);
-    let location = resp
-        .headers()
-        .get("Location")
-        .and_then(|value| value.to_str().ok())
-        .expect("OIDC error redirect location should exist");
-    assert!(location.starts_with("http://localhost:8080/login?external_auth=error"));
-    assert!(common::extract_cookie(resp, "aster_access").is_none());
-    assert!(common::extract_cookie(resp, "aster_refresh").is_none());
-}
-
-fn oidc_email_required_flow<B>(resp: &ServiceResponse<B>) -> String {
-    assert_eq!(resp.status(), 302);
-    let location = resp
-        .headers()
-        .get("Location")
-        .and_then(|value| value.to_str().ok())
-        .expect("OIDC email required redirect location should exist");
-    assert!(location.starts_with("http://localhost:8080/login?external_auth=email_required"));
-    assert!(common::extract_cookie(resp, "aster_access").is_none());
-    assert!(common::extract_cookie(resp, "aster_refresh").is_none());
-    let parsed = reqwest::Url::parse(location).expect("redirect location should parse");
-    parsed
-        .query_pairs()
-        .find(|(key, _)| key == "flow")
-        .map(|(_, value)| value.into_owned())
-        .expect("email required redirect should include flow token")
-}
-
-async fn start_oidc_email_verification<S, B, E>(
-    app: &S,
-    flow_token: &str,
-    email: &str,
-) -> ServiceResponse<B>
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let req = test::TestRequest::post()
-        .uri("/api/v1/auth/external-auth/email-verification/start")
-        .insert_header(("Origin", TEST_BROWSER_ORIGIN))
-        .set_json(serde_json::json!({
-            "flow_token": flow_token,
-            "email": email
-        }))
-        .to_request();
-    test::call_service(app, req).await
-}
-
-async fn assert_start_oidc_email_verification_ok<S, B, E>(app: &S, flow_token: &str, email: &str)
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let resp = start_oidc_email_verification(app, flow_token, email).await;
-    assert_eq!(resp.status(), 200);
-    let body: Value = test::read_body_json(resp).await;
-    assert_eq!(
-        body["data"]["message"],
-        "external auth email verification email sent"
-    );
-}
-
-async fn confirm_oidc_email_verification<S, B, E>(app: &S, token: &str) -> ServiceResponse<B>
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let req = test::TestRequest::get()
-        .uri(&format!(
-            "/api/v1/auth/external-auth/email-verification/confirm?token={}",
-            urlencoding::encode(token)
-        ))
-        .peer_addr("127.0.0.1:12345".parse().unwrap())
-        .to_request();
-    test::call_service(app, req).await
-}
-
-async fn link_oidc_with_password<S, B, E>(
-    app: &S,
-    flow_token: &str,
-    identifier: &str,
-    password: &str,
-) -> ServiceResponse<B>
-where
-    S: actix_web::dev::Service<
-            actix_http::Request,
-            Response = actix_web::dev::ServiceResponse<B>,
-            Error = E,
-        >,
-    B: MessageBody,
-    E: std::fmt::Debug,
-{
-    let req = test::TestRequest::post()
-        .uri("/api/v1/auth/external-auth/password-link")
-        .insert_header(("Origin", TEST_BROWSER_ORIGIN))
-        .peer_addr("127.0.0.1:12345".parse().unwrap())
-        .set_json(serde_json::json!({
-            "flow_token": flow_token,
-            "identifier": identifier,
-            "password": password
-        }))
-        .to_request();
-    test::call_service(app, req).await
-}
-
-async fn latest_oidc_email_verification_token(
-    state: &aster_drive::runtime::PrimaryAppState,
-) -> String {
-    common::flush_mail_outbox(state).await;
-    let memory_sender = aster_drive::services::mail_service::memory_sender_ref(&state.mail_sender)
-        .expect("memory mail sender should be available in tests");
-    let message = memory_sender
-        .last_message()
-        .expect("OIDC email verification email should be sent");
-    common::extract_token_from_mail_message(
-        &message,
-        "/api/v1/auth/external-auth/email-verification/confirm?token=",
-    )
-    .expect("OIDC email verification token should be present")
-}
-
-async fn disable_user(state: &aster_drive::runtime::PrimaryAppState, user_id: i64) {
-    let user = user::Entity::find_by_id(user_id)
-        .one(&state.db)
-        .await
-        .expect("user should query")
-        .expect("user should exist");
-    let mut active = user.into_active_model();
-    active.status = Set(aster_drive::types::UserStatus::Disabled);
-    active.update(&state.db).await.expect("user should update");
-}
-
-fn reserve_localhost_port() -> (u16, std::net::TcpListener) {
-    let listener = std::net::TcpListener::bind(("127.0.0.1", 0))
-        .expect("free localhost port should be reserved");
-    let port = listener
-        .local_addr()
-        .expect("reserved listener address should exist")
-        .port();
-    (port, listener)
-}
-
-fn dex_config(issuer: &str, provider_key: &str) -> String {
-    format!(
-        r#"issuer: {issuer}
-storage:
-  type: memory
-web:
-  http: 0.0.0.0:5556
-oauth2:
-  skipApprovalScreen: true
-staticClients:
-  - id: {TEST_CLIENT_ID}
-    redirectURIs:
-      - {TEST_BROWSER_ORIGIN}/api/v1/auth/external-auth/oidc/{provider_key}/callback
-    name: AsterDrive Test
-    secret: {DEX_TEST_CLIENT_SECRET}
-enablePasswordDB: true
-staticPasswords:
-  - email: "{DEX_TEST_USER_EMAIL}"
-    hash: "$2a$10$2b2cU8CPhOTaGrs1HRQuAueS7JTT5ZHsHSzYiFPm1leZck7Mc8T4W"
-    username: "dex-user"
-    userID: "dex-user-id"
-"#
-    )
-}
-
-async fn wait_for_dex_discovery(issuer: &str) {
-    let discovery_url = format!("{issuer}/.well-known/openid-configuration");
-    let client = reqwest::Client::builder()
-        .redirect(reqwest::redirect::Policy::none())
-        .timeout(StdDuration::from_secs(2))
-        .build()
-        .expect("reqwest client should build");
-    let deadline = tokio::time::Instant::now() + StdDuration::from_secs(30);
-    let mut last_error: Option<String>;
-    loop {
-        last_error = match client.get(&discovery_url).send().await {
-            Ok(resp) if resp.status().is_success() => return,
-            Ok(resp) => Some(format!("HTTP {}", resp.status())),
-            Err(err) => Some(err.to_string()),
-        };
-        assert!(
-            tokio::time::Instant::now() < deadline,
-            "timed out waiting for Dex discovery at {discovery_url}: {}",
-            last_error.unwrap_or_else(|| "unknown error".to_string())
-        );
-        tokio::time::sleep(StdDuration::from_millis(250)).await;
-    }
-}
-
-fn absolute_location(base: &reqwest::Url, location: &str) -> reqwest::Url {
-    reqwest::Url::parse(location)
-        .or_else(|_| base.join(location))
-        .expect("redirect location should be a valid URL")
-}
-
-async fn request_dex_redirect(
-    client: &reqwest::Client,
-    url: reqwest::Url,
-) -> (reqwest::Url, Option<reqwest::Url>) {
-    let resp = client
-        .get(url.clone())
-        .send()
-        .await
-        .expect("Dex redirect GET should succeed");
-    if resp.status().is_redirection() {
-        let location = resp
-            .headers()
-            .get(reqwest::header::LOCATION)
-            .and_then(|value| value.to_str().ok())
-            .expect("Dex redirect should include Location");
-        return (url.clone(), Some(absolute_location(&url, location)));
-    }
-    panic!(
-        "Dex GET {url} returned non-redirect status {}",
-        resp.status()
-    );
-}
-
-async fn complete_dex_password_login(
-    issuer: &str,
-    provider_key: &str,
-    authorization_url: &str,
-) -> String {
-    let client = reqwest::Client::builder()
-        .redirect(reqwest::redirect::Policy::none())
-        .timeout(StdDuration::from_secs(10))
-        .build()
-        .expect("reqwest client should build");
-    let issuer_url = reqwest::Url::parse(issuer).expect("Dex issuer URL should parse");
-    let mut next =
-        reqwest::Url::parse(authorization_url).expect("OIDC authorization URL should parse");
-    let mut login_url = None;
-
-    for _ in 0..6 {
-        let (_, redirect) = request_dex_redirect(&client, next.clone()).await;
-        let redirect = redirect.expect("Dex should keep redirecting until the login form");
-        if redirect.path().ends_with("/auth/local/login") {
-            login_url = Some(redirect);
-            break;
-        }
-        assert_eq!(
-            redirect.domain(),
-            issuer_url.domain(),
-            "Dex should not redirect to another host before login"
-        );
-        next = redirect;
-    }
-
-    let login_url = login_url.expect("Dex password login URL should be reached");
-    let form = format!(
-        "login={}&password={}",
-        urlencoding::encode(DEX_TEST_USER_EMAIL),
-        urlencoding::encode("password")
-    );
-    let resp = client
-        .post(login_url.clone())
-        .header(
-            reqwest::header::CONTENT_TYPE,
-            "application/x-www-form-urlencoded",
-        )
-        .body(form)
-        .send()
-        .await
-        .expect("Dex password POST should succeed");
-    assert!(
-        resp.status().is_redirection(),
-        "Dex password POST should redirect, got {}",
-        resp.status()
-    );
-    let location = resp
-        .headers()
-        .get(reqwest::header::LOCATION)
-        .and_then(|value| value.to_str().ok())
-        .expect("Dex login redirect should include Location");
-    let redirect = absolute_location(&login_url, location);
-    let expected_callback =
-        format!("{TEST_BROWSER_ORIGIN}/api/v1/auth/external-auth/oidc/{provider_key}/callback");
-    assert_eq!(
-        redirect.as_str().split('?').next(),
-        Some(expected_callback.as_str()),
-        "Dex should redirect back to AsterDrive callback"
-    );
-    redirect.to_string()
-}
 
 #[actix_web::test]
 async fn admin_provider_api_masks_secret_and_public_list_only_shows_enabled() {
